Adding query infrastructure and indexes
Iterator strategy is a bit of dead end because we really want a smarter
scanning strategy based on the filter, rather than to just take every
directory entry recursively.

As I type that above, I realize that the iterator may still work, you'd
just initialize it with the filter perhaps. This would move all the
implementation to the iterator? We still then need some container to
feed in what it iterates over. For DI.